Struct run_make_support::command::Command

pub struct Command {
    cmd: Command,
    stdin: Option<Box<[u8]>>,
    drop_bomb: DropBomb,
}

This is a custom command wrapper that simplifies working with commands and makes it easier to ensure that we check the exit status of executed processes.

A Command must be executed

A Command is armed by a DropBomb on construction to enforce that it will be executed. If a Command is constructed but never executed, the drop bomb will explode and cause the test to panic. Execution methods run and run_fail will defuse the drop bomb. A test containing constructed but never executed commands is dangerous because it can give a false sense of confidence.

Fields
cmd: Commandstdin: Option<Box<[u8]>>drop_bomb: DropBomb
Implementations

impl Command

pub fn new<P: AsRef<OsStr>>(program: P) -> Self

pub fn set_stdin(&mut self, stdin: Box<[u8]>)

pub fn env<K, V>(&mut self, key: K, value: V) -> &mut Self
where K: AsRef<OsStr>, V: AsRef<OsStr>,

Specify an environment variable.

pub fn env_remove<K>(&mut self, key: K) -> &mut Self
where K: AsRef<OsStr>,

Remove an environmental variable.

pub fn arg<S>(&mut self, arg: S) -> &mut Self
where S: AsRef<OsStr>,

Generic command argument provider. Prefer specific helper methods if possible. Note that for some executables, arguments might be platform specific. For C/C++ compilers, arguments might be platform and compiler specific.

pub fn args<S>(&mut self, args: &[S]) -> &mut Self
where S: AsRef<OsStr>,

Generic command arguments provider. Prefer specific helper methods if possible. Note that for some executables, arguments might be platform specific. For C/C++ compilers, arguments might be platform and compiler specific.

pub fn inspect<I>(&mut self, inspector: I) -> &mut Self
where I: FnOnce(&StdCommand),

Inspect what the underlying std::process::Command is up to the current construction.

pub fn current_dir<P: AsRef<Path>>(&mut self, path: P) -> &mut Self

Set the path where the command will be run.

pub fn run(&mut self) -> CompletedProcess

Run the constructed command and assert that it is successfully run.

pub fn run_fail(&mut self) -> CompletedProcess

Run the constructed command and assert that it does not successfully run.

pub fn run_unchecked(&mut self) -> CompletedProcess

Run the command but do not check its exit status. Only use if you explicitly don’t care about the exit status. Prefer to use Self::run and Self::run_fail whenever possible.

fn command_output(&mut self) -> CompletedProcess