Module run_make_support::external_deps

This module contains external tool dependencies that we assume are available in the environment, such as cc or python.

Modules

c_build

cc

clang

cygpath 🔒

htmldocck

llvm

python

rustc

rustdoc