Struct run_make_support::external_deps::clang::Clang

pub struct Clang {
    cmd: Command,
}

A clang invocation builder.

Fields
cmd: Command
Implementations

impl Clang

pub fn env<K, V>(&mut self, key: K, value: V) -> &mut Self
where K: AsRef<OsStr>, V: AsRef<OsStr>,

Specify an environment variable.

pub fn env_remove<K>(&mut self, key: K) -> &mut Self
where K: AsRef<OsStr>,

Remove an environmental variable.

pub fn arg<S>(&mut self, arg: S) -> &mut Self
where S: AsRef<OsStr>,

Generic command argument provider. Prefer specific helper methods if possible. Note that for some executables, arguments might be platform specific. For C/C++ compilers, arguments might be platform and compiler specific.

pub fn args<V, S>(&mut self, args: V) -> &mut Self
where V: AsRef<[S]>, S: AsRef<OsStr>,

Generic command arguments provider. Prefer specific helper methods if possible. Note that for some executables, arguments might be platform specific. For C/C++ compilers, arguments might be platform and compiler specific.

pub fn inspect<I>(&mut self, inspector: I) -> &mut Self
where I: FnOnce(&Command),

Inspect what the underlying Command is up to the current construction.

pub fn run(&mut self) -> CompletedProcess

Run the constructed command and assert that it is successfully run.

pub fn run_fail(&mut self) -> CompletedProcess

Run the constructed command and assert that it does not successfully run.

pub fn run_unchecked(&mut self) -> CompletedProcess

Run the command but do not check its exit status. Only use if you explicitly don’t care about the exit status. Prefer to use Self::run and Self::run_fail whenever possible.

pub fn current_dir<P: AsRef<Path>>(&mut self, path: P) -> &mut Self

Set the path where the command will be run.

impl Clang

pub fn new() -> Self

Construct a new clang invocation. clang is not always available for all targets.

pub fn input<P: AsRef<Path>>(&mut self, path: P) -> &mut Self

Provide an input file.

pub fn out_exe(&mut self, name: &str) -> &mut Self

Specify the name of the executable. The executable will be placed under the current directory and the extension will be determined by bin_name.

pub fn target(&mut self, target_triple: &str) -> &mut Self

Specify which target triple clang should target.

pub fn no_stdlib(&mut self) -> &mut Self

Pass -nostdlib to disable linking the C standard library.

pub fn arch(&mut self, arch: &str) -> &mut Self

Specify architecture.

pub fn lto(&mut self, lto: &str) -> &mut Self

Specify LTO settings.

pub fn use_ld(&mut self, ld: &str) -> &mut Self

Specify which ld to use.