Struct run_make_support::external_deps::llvm::LlvmAr

pub struct LlvmAr {
    cmd: Command,
}

A llvm-ar invocation builder.

Fields
cmd: Command
Implementations

impl LlvmAr

pub fn env<K, V>(&mut self, key: K, value: V) -> &mut Self
where K: AsRef<OsStr>, V: AsRef<OsStr>,

Specify an environment variable.

pub fn env_remove<K>(&mut self, key: K) -> &mut Self
where K: AsRef<OsStr>,

Remove an environmental variable.

pub fn arg<S>(&mut self, arg: S) -> &mut Self
where S: AsRef<OsStr>,

Generic command argument provider. Prefer specific helper methods if possible. Note that for some executables, arguments might be platform specific. For C/C++ compilers, arguments might be platform and compiler specific.

pub fn args<V, S>(&mut self, args: V) -> &mut Self
where V: AsRef<[S]>, S: AsRef<OsStr>,

Generic command arguments provider. Prefer specific helper methods if possible. Note that for some executables, arguments might be platform specific. For C/C++ compilers, arguments might be platform and compiler specific.

pub fn inspect<I>(&mut self, inspector: I) -> &mut Self
where I: FnOnce(&Command),

Inspect what the underlying Command is up to the current construction.

pub fn run(&mut self) -> CompletedProcess

Run the constructed command and assert that it is successfully run.

pub fn run_fail(&mut self) -> CompletedProcess

Run the constructed command and assert that it does not successfully run.

pub fn run_unchecked(&mut self) -> CompletedProcess

Run the command but do not check its exit status. Only use if you explicitly don’t care about the exit status. Prefer to use Self::run and Self::run_fail whenever possible.

pub fn current_dir<P: AsRef<Path>>(&mut self, path: P) -> &mut Self

Set the path where the command will be run.

impl LlvmAr

pub fn new() -> Self

Construct a new llvm-ar invocation. This assumes that llvm-ar is available at $LLVM_BIN_DIR/llvm-ar.

pub fn obj_to_ar(&mut self) -> &mut Self

Automatically pass the commonly used arguments rcus, used for combining one or more input object files into one output static library file.

pub fn output_input( &mut self, out: impl AsRef<Path>, input: impl AsRef<Path>, ) -> &mut Self

Provide an output, then an input file. Bundled in one function, as llvm-ar has no “–output”-style flag.