run_make_support::external_deps::rustc

Struct Rustc

pub struct Rustc {
    cmd: Command,
    target: Option<String>,
}

A rustc invocation builder.

Fields
cmd: Commandtarget: Option<String>
Implementations

impl Rustc

pub fn into_raw_command(self) -> Command

In very rare circumstances, you may need a e.g. bare_rustc() or bare_rustdoc() with host runtime libs configured, but want the underlying raw std::process::Command (e.g. for manipulating pipes or whatever). This function will consume the command wrapper and extract the underlying std::process::Command.

Caution: this will mean that you can no longer use the convenience methods on the command wrapper. Use as a last resort.

pub fn env<K, V>(&mut self, key: K, value: V) -> &mut Self
where K: AsRef<OsStr>, V: AsRef<OsStr>,

Specify an environment variable.

pub fn env_remove<K>(&mut self, key: K) -> &mut Self
where K: AsRef<OsStr>,

Remove an environmental variable.

pub fn arg<S>(&mut self, arg: S) -> &mut Self
where S: AsRef<OsStr>,

Generic command argument provider. Prefer specific helper methods if possible. Note that for some executables, arguments might be platform specific. For C/C++ compilers, arguments might be platform and compiler specific.

pub fn args<V, S>(&mut self, args: V) -> &mut Self
where V: AsRef<[S]>, S: AsRef<OsStr>,

Generic command arguments provider. Prefer specific helper methods if possible. Note that for some executables, arguments might be platform specific. For C/C++ compilers, arguments might be platform and compiler specific.

pub fn stdin<T: Into<Stdio>>(&mut self, cfg: T) -> &mut Self

Configuration for the child process’s standard input (stdin) handle.

See std::process::Command::stdin.

pub fn stdout<T: Into<Stdio>>(&mut self, cfg: T) -> &mut Self

Configuration for the child process’s standard output (stdout) handle.

See std::process::Command::stdout.

pub fn stderr<T: Into<Stdio>>(&mut self, cfg: T) -> &mut Self

Configuration for the child process’s standard error (stderr) handle.

See std::process::Command::stderr.

pub fn inspect<I>(&mut self, inspector: I) -> &mut Self
where I: FnOnce(&Command),

Inspect what the underlying Command is up to the current construction.

pub fn set_aux_fd<F: Into<OwnedFd>>( &mut self, new_fd: RawFd, fd: F, ) -> &mut Self

Set an auxiliary stream passed to the process, besides the stdio streams.

pub fn run(&mut self) -> CompletedProcess

Run the constructed command and assert that it is successfully run.

pub fn run_fail(&mut self) -> CompletedProcess

Run the constructed command and assert that it does not successfully run.

pub fn run_unchecked(&mut self) -> CompletedProcess

Run the command but do not check its exit status. Only use if you explicitly don’t care about the exit status. Prefer to use Self::run and Self::run_fail whenever possible.

pub fn current_dir<P: AsRef<Path>>(&mut self, path: P) -> &mut Self

Set the path where the command will be run.

impl Rustc

pub fn new() -> Self

Construct a new rustc invocation. This will automatically set the library search path as -L cwd(), configure the compilation target and enable dynamic linkage by default on musl hosts. Use bare_rustc to avoid this.

pub fn bare() -> Self

Construct a bare rustc invocation with no flags set.

pub fn cfg(&mut self, s: &str) -> &mut Self

Configure the compilation environment.

pub fn opt(&mut self) -> &mut Self

Specify default optimization level -O (alias for -C opt-level=3).

pub fn opt_level(&mut self, option: &str) -> &mut Self

Specify a specific optimization level.

pub fn metadata(&mut self, meta: &str) -> &mut Self

Incorporate a hashed string to mangled symbols.

pub fn extra_filename(&mut self, suffix: &str) -> &mut Self

Add a suffix in each output filename.

pub fn emit<S: AsRef<str>>(&mut self, kinds: S) -> &mut Self

Specify type(s) of output files to generate.

pub fn extern_<P: AsRef<Path>>( &mut self, crate_name: &str, path: P, ) -> &mut Self

Specify where an external library is located.

pub fn remap_path_prefix<P: AsRef<Path>, P2: AsRef<Path>>( &mut self, from: P, to: P2, ) -> &mut Self

Remap source path prefixes in all output.

pub fn input<P: AsRef<Path>>(&mut self, path: P) -> &mut Self

Specify path to the input file.

pub fn set_backtrace_level<R: AsRef<OsStr>>(&mut self, level: R) -> &mut Self

pub fn output<P: AsRef<Path>>(&mut self, path: P) -> &mut Self

Specify path to the output file. Equivalent to `-o`` in rustc.

pub fn out_dir<P: AsRef<Path>>(&mut self, path: P) -> &mut Self

Specify path to the output directory. Equivalent to `–out-dir`` in rustc.

pub fn lto(&mut self, option: &str) -> &mut Self

This flag enables LTO in the specified form.

pub fn linker_plugin_lto(&mut self, option: &str) -> &mut Self

This flag defers LTO optimizations to the linker.

pub fn panic(&mut self, option: &str) -> &mut Self

Specify what happens when the code panics.

pub fn codegen_units(&mut self, units: usize) -> &mut Self

Specify number of codegen units

pub fn incremental<P: AsRef<Path>>(&mut self, path: P) -> &mut Self

Specify directory path used for incremental cache

pub fn profile_generate<P: AsRef<Path>>(&mut self, path: P) -> &mut Self

Specify directory path used for profile generation

pub fn profile_use<P: AsRef<Path>>(&mut self, path: P) -> &mut Self

Specify directory path used for profile usage

pub fn symbol_mangling_version(&mut self, option: &str) -> &mut Self

Specify option of -C symbol-mangling-version.

pub fn prefer_dynamic(&mut self) -> &mut Self

Specify -C prefer-dynamic.

pub fn error_format(&mut self, format: &str) -> &mut Self

Specify error format to use

pub fn json(&mut self, items: &str) -> &mut Self

Specify json messages printed by the compiler

pub fn ui_testing(&mut self) -> &mut Self

Normalize the line number in the stderr output

pub fn target<S: AsRef<str>>(&mut self, target: S) -> &mut Self

Specify the target triple, or a path to a custom target json spec file.

pub fn target_cpu<S: AsRef<str>>(&mut self, target_cpu: S) -> &mut Self

Specify the target CPU.

pub fn crate_type(&mut self, crate_type: &str) -> &mut Self

Specify the crate type.

pub fn library_search_path<P: AsRef<Path>>(&mut self, path: P) -> &mut Self

Add a directory to the library search path. Equivalent to -L in rustc.

pub fn specific_library_search_path<P: AsRef<Path>>( &mut self, kind: &str, path: P, ) -> &mut Self

Add a directory to the library search path with a restriction, where kind is a dependency type. Equivalent to -L KIND=PATH in rustc.

pub fn sysroot<P: AsRef<Path>>(&mut self, path: P) -> &mut Self

Override the system root. Equivalent to --sysroot in rustc.

pub fn edition(&mut self, edition: &str) -> &mut Self

Specify the edition year.

pub fn print(&mut self, request: &str) -> &mut Self

Specify the print request.

pub fn link_arg(&mut self, link_arg: &str) -> &mut Self

Add an extra argument to the linker invocation, via -Clink-arg.

pub fn link_args(&mut self, link_args: &str) -> &mut Self

Add multiple extra arguments to the linker invocation, via -Clink-args.

pub fn stdin_buf<I: AsRef<[u8]>>(&mut self, input: I) -> &mut Self

Specify a stdin input buffer.

pub fn crate_name<S: AsRef<OsStr>>(&mut self, name: S) -> &mut Self

Specify the crate name.

pub fn linker(&mut self, linker: &str) -> &mut Self

Specify the linker

pub fn linker_flavor(&mut self, linker_flavor: &str) -> &mut Self

Specify the linker flavor

pub fn debuginfo(&mut self, level: &str) -> &mut Self

Specify -C debuginfo=....

pub fn split_debuginfo(&mut self, split_kind: &str) -> &mut Self

Specify -C split-debuginfo={packed,unpacked,off}.

pub fn verbose(&mut self) -> &mut Self

Pass the --verbose flag.

pub fn extra_rs_cxx_flags(&mut self) -> &mut Self

EXTRARSCXXFLAGS