Module run_make_support::fs_wrapper

Functions

copy

A wrapper around std::fs::copy which includes the file path in the panic message.

create_dir

A wrapper around std::fs::create_dir which includes the file path in the panic message..

create_dir_all

A wrapper around std::fs::create_dir_all which includes the file path in the panic message..

create_file

A wrapper around std::fs::File::create which includes the file path in the panic message..

metadata

A wrapper around std::fs::metadata which includes the file path in the panic message..

read

A wrapper around std::fs::read which includes the file path in the panic message..

read_dir

A wrapper around std::fs::read_dir which includes the file path in the panic message..

read_to_string

A wrapper around std::fs::read_to_string which includes the file path in the panic message..

remove_dir_all

A wrapper around std::fs::remove_dir_all which includes the file path in the panic message..

remove_file

A wrapper around std::fs::remove_file which includes the file path in the panic message..

rename

A wrapper around std::fs::rename which includes the file path in the panic message.

set_permissions

A wrapper around std::fs::set_permissions which includes the file path in the panic message.

write

A wrapper around std::fs::write which includes the file path in the panic message..