Module run_make_support::rfs

std::fs wrappers and assorted filesystem-related helpers. Public to tests as rfs to not be confused with std::fs.

Functions

copy

A wrapper around std::fs::copy which includes the file path in the panic message.

copy_dir_all

Copy a directory into another.

create_dir

A wrapper around std::fs::create_dir which includes the file path in the panic message.

create_dir_all

A wrapper around std::fs::create_dir_all which includes the file path in the panic message.

create_file

A wrapper around std::fs::File::create which includes the file path in the panic message.

create_symlink

Creates a new symlink to a path on the filesystem, adjusting for Windows or Unix.

metadata

A wrapper around std::fs::metadata which includes the file path in the panic message.

read

A wrapper around std::fs::read which includes the file path in the panic message.

read_dir

A wrapper around std::fs::read_dir which includes the file path in the panic message.

read_dir_entries

Helper for reading entries in a given directory.

read_to_string

A wrapper around std::fs::read_to_string which includes the file path in the panic message.

remove_dir_all

A wrapper around std::fs::remove_dir_all which includes the file path in the panic message.

remove_file

A wrapper around std::fs::remove_file which includes the file path in the panic message.

rename

A wrapper around std::fs::rename which includes the file path in the panic message.

set_permissions

A wrapper around std::fs::set_permissions which includes the file path in the panic message.

write

A wrapper around std::fs::write which includes the file path in the panic message.