run_make_support::rfs

Function metadata

pub fn metadata<P: AsRef<Path>>(path: P) -> Metadata

A wrapper around std::fs::metadata which includes the file path in the panic message. Note that this will traverse symlinks and will return metadata about the target file. Use symlink_metadata if you don’t want to traverse symlinks.

See std::fs::metadata docs for more details.