run_make_support
Module run

Functions

cmd

Create a new custom Command. This should be preferred to creating std::process::Command directly.

run

Run a built binary and make sure it succeeds.

run_common 🔒

run_fail

Run a built binary and make sure it fails.

run_with_args

Run a built binary with one or more argument(s) and make sure it succeeds.