Struct run_make_support::rustc::Rustc

pub struct Rustc {
    cmd: Command,
}

A rustc invocation builder.

Fields
cmd: Command
Implementations

impl Rustc

pub fn env<K, V>(&mut self, key: K, value: V) -> &mut Self
where K: AsRef<OsStr>, V: AsRef<OsStr>,

Specify an environment variable.

pub fn env_remove<K>(&mut self, key: K) -> &mut Self
where K: AsRef<OsStr>,

Remove an environmental variable.

pub fn arg<S>(&mut self, arg: S) -> &mut Self
where S: AsRef<OsStr>,

Generic command argument provider. Prefer specific helper methods if possible. Note that for some executables, arguments might be platform specific. For C/C++ compilers, arguments might be platform and compiler specific.

pub fn args<S>(&mut self, args: &[S]) -> &mut Self
where S: AsRef<OsStr>,

Generic command arguments provider. Prefer specific helper methods if possible. Note that for some executables, arguments might be platform specific. For C/C++ compilers, arguments might be platform and compiler specific.

pub fn inspect<I>(&mut self, inspector: I) -> &mut Self
where I: FnOnce(&Command),

Inspect what the underlying Command is up to the current construction.

pub fn run(&mut self) -> CompletedProcess

Run the constructed command and assert that it is successfully run.

pub fn run_fail(&mut self) -> CompletedProcess

Run the constructed command and assert that it does not successfully run.

pub fn run_unchecked(&mut self) -> CompletedProcess

Run the command but do not check its exit status. Only use if you explicitly don’t care about the exit status. Prefer to use Self::run and Self::run_fail whenever possible.

pub fn current_dir<P: AsRef<Path>>(&mut self, path: P) -> &mut Self

Set the path where the command will be run.

impl Rustc

pub fn new() -> Self

Construct a new rustc invocation.

pub fn new_aux_build() -> Self

Construct a new rustc invocation with aux_build preset (setting --crate-type=lib).

pub fn cfg(&mut self, s: &str) -> &mut Self

Configure the compilation environment.

pub fn opt(&mut self) -> &mut Self

Specify default optimization level -O (alias for -C opt-level=2).

pub fn opt_level(&mut self, option: &str) -> &mut Self

Specify a specific optimization level.

pub fn extra_filename(&mut self, suffix: &str) -> &mut Self

Add a suffix in each output filename.

pub fn emit(&mut self, kinds: &str) -> &mut Self

Specify type(s) of output files to generate.

pub fn extern_<P: AsRef<Path>>( &mut self, crate_name: &str, path: P, ) -> &mut Self

Specify where an external library is located.

pub fn input<P: AsRef<Path>>(&mut self, path: P) -> &mut Self

Specify path to the input file.

pub fn set_backtrace_level<R: AsRef<OsStr>>(&mut self, level: R) -> &mut Self

pub fn output<P: AsRef<Path>>(&mut self, path: P) -> &mut Self

Specify path to the output file. Equivalent to `-o`` in rustc.

pub fn out_dir<P: AsRef<Path>>(&mut self, path: P) -> &mut Self

Specify path to the output directory. Equivalent to `–out-dir`` in rustc.

pub fn linker_plugin_lto(&mut self, option: &str) -> &mut Self

This flag defers LTO optimizations to the linker.

pub fn panic(&mut self, option: &str) -> &mut Self

Specify what happens when the code panics.

pub fn codegen_units(&mut self, units: usize) -> &mut Self

Specify number of codegen units

pub fn incremental<P: AsRef<Path>>(&mut self, path: P) -> &mut Self

Specify directory path used for incremental cache

pub fn profile_generate<P: AsRef<Path>>(&mut self, path: P) -> &mut Self

Specify directory path used for profile generation

pub fn profile_use<P: AsRef<Path>>(&mut self, path: P) -> &mut Self

Specify directory path used for profile usage

pub fn error_format(&mut self, format: &str) -> &mut Self

Specify error format to use

pub fn json(&mut self, items: &str) -> &mut Self

Specify json messages printed by the compiler

pub fn target(&mut self, target: &str) -> &mut Self

Specify the target triple, or a path to a custom target json spec file.

pub fn crate_type(&mut self, crate_type: &str) -> &mut Self

Specify the crate type.

pub fn library_search_path<P: AsRef<Path>>(&mut self, path: P) -> &mut Self

Add a directory to the library search path. Equivalent to -L in rustc.

pub fn sysroot<P: AsRef<Path>>(&mut self, path: P) -> &mut Self

Override the system root. Equivalent to --sysroot in rustc.

pub fn edition(&mut self, edition: &str) -> &mut Self

Specify the edition year.

pub fn print(&mut self, request: &str) -> &mut Self

Specify the print request.

pub fn link_arg(&mut self, link_arg: &str) -> &mut Self

Add an extra argument to the linker invocation, via -Clink-arg.

pub fn stdin<I: AsRef<[u8]>>(&mut self, input: I) -> &mut Self

Specify a stdin input

pub fn crate_name<S: AsRef<OsStr>>(&mut self, name: S) -> &mut Self

Specify the crate name.

pub fn linker(&mut self, linker: &str) -> &mut Self

Specify the linker