Struct run_make_support::rustdoc::Rustdoc

pub struct Rustdoc {
    cmd: Command,
}

Fields
cmd: Command
Implementations

impl Rustdoc

pub fn env<K, V>(&mut self, key: K, value: V) -> &mut Self
where K: AsRef<OsStr>, V: AsRef<OsStr>,

Specify an environment variable.

pub fn env_remove<K>(&mut self, key: K) -> &mut Self
where K: AsRef<OsStr>,

Remove an environmental variable.

pub fn arg<S>(&mut self, arg: S) -> &mut Self
where S: AsRef<OsStr>,

Generic command argument provider. Prefer specific helper methods if possible. Note that for some executables, arguments might be platform specific. For C/C++ compilers, arguments might be platform and compiler specific.

pub fn args<S>(&mut self, args: &[S]) -> &mut Self
where S: AsRef<OsStr>,

Generic command arguments provider. Prefer specific helper methods if possible. Note that for some executables, arguments might be platform specific. For C/C++ compilers, arguments might be platform and compiler specific.

pub fn inspect<I>(&mut self, inspector: I) -> &mut Self
where I: FnOnce(&Command),

Inspect what the underlying Command is up to the current construction.

pub fn run(&mut self) -> CompletedProcess

Run the constructed command and assert that it is successfully run.

pub fn run_fail(&mut self) -> CompletedProcess

Run the constructed command and assert that it does not successfully run.

pub fn run_unchecked(&mut self) -> CompletedProcess

Run the command but do not check its exit status. Only use if you explicitly don’t care about the exit status. Prefer to use Self::run and Self::run_fail whenever possible.

pub fn current_dir<P: AsRef<Path>>(&mut self, path: P) -> &mut Self

Set the path where the command will be run.

impl Rustdoc

pub fn bare() -> Self

Construct a bare rustdoc invocation.

pub fn new() -> Self

Construct a rustdoc invocation with -L $(TARGET_RPATH_DIR) set.

pub fn extern_<P: AsRef<Path>>( &mut self, crate_name: &str, path: P, ) -> &mut Self

Specify where an external library is located.

pub fn input<P: AsRef<Path>>(&mut self, path: P) -> &mut Self

Specify path to the input file.

pub fn output<P: AsRef<Path>>(&mut self, path: P) -> &mut Self

Specify path to the output folder.

pub fn out_dir<P: AsRef<Path>>(&mut self, path: P) -> &mut Self

Specify output directory.

pub fn arg_file<P: AsRef<Path>>(&mut self, path: P) -> &mut Self

Given a path, pass @{path} to rustdoc as an arg file.

pub fn stdin<I: AsRef<[u8]>>(&mut self, input: I) -> &mut Self

Specify a stdin input

pub fn edition(&mut self, edition: &str) -> &mut Self

Specify the edition year.

pub fn target(&mut self, target: &str) -> &mut Self

Specify the target triple, or a path to a custom target json spec file.

pub fn crate_type(&mut self, crate_type: &str) -> &mut Self

Specify the crate type.

pub fn crate_name<S: AsRef<OsStr>>(&mut self, name: S) -> &mut Self

Specify the crate name.

pub fn library_search_path<P: AsRef<Path>>(&mut self, path: P) -> &mut Self

Add a directory to the library search path. It corresponds to the -L rustdoc option.

pub fn output_format(&mut self, format: &str) -> &mut Self

Specify the output format.