run_make_support
Module scoped_run

Collection of helpers that try to maintain certain properties while running a test closure.

Functions

run_in_tmpdir

This function is designed for running commands in a temporary directory that is cleared after the function ends.

test_while_readonly

Ensure that the path P is read-only while the test runs, and restore original permissions at the end so compiletest can clean up. This will panic on Windows if the path is a directory (as it would otherwise do nothing)