Module run_make_support::string

Functions

count_regex_matches_in_files_with_extension

Gathers all files in the current working directory that have the extension ext, and counts the number of lines within that contain a match with the regex pattern re.

invalid_utf8_contains

Read the contents of a file that cannot simply be read by read_to_string, due to invalid UTF-8 data, then assert that it contains expected.

invalid_utf8_not_contains

Read the contents of a file that cannot simply be read by read_to_string, due to invalid UTF-8 data, then assert that it does not contain expected.