run_make_support::symbols

Function object_contains_all_symbol_substring

pub fn object_contains_all_symbol_substring<'s, P, S>(
    path: P,
    substrings: &'s [S],
) -> ContainsAllSymbolSubstringsOutcome<'s>
where
    P: AsRef<Path>,
    S: AsRef<str>,

Check an object file’s symbols for presence of all of provided substrings. That is, if an object file contains symbols ["hello", "goodbye", "world"], it will be matched against a list of substrings of ["he", "go"]. In this case, he is a substring of hello, and go is a substring of goodbye, so each of substrings was found.

Returns true if all substrings were present in the names of symbols for the given object file (as substrings of symbol names).

Panics if path is not a valid object file readable by the current user or if path cannot be parsed as a recognized object file.

Platform-specific behavior

See object_contains_any_symbol_substring.