run_make_support::symbols

Function object_contains_any_symbol

pub fn object_contains_any_symbol<P, S>(
    path: P,
    candidate_symbols: &[S],
) -> bool
where
    P: AsRef<Path>,
    S: AsRef<str>,

Check an object file’s symbols for any exact matches against those provided in candidate_symbols.

Returns true if any of the symbols found in the object file at path contain an exact match against those listed in candidate_symbols. Take care to account for (1) platform differences and (2) calling convention and symbol decorations differences.

Panics if path is not a valid object file readable by the current user or if path cannot be parsed as a recognized object file.

Platform-specific behavior

See object_contains_any_symbol_substring.