Crate run_make_support

run-make-support is a support library for run-make tests. It provides command wrappers and convenience utility functions to help test writers reduce duplication. The support library notably is built via cargo: this means that if your test wants some non-trivial utility, such as object or wasmparser, they can be re-exported and be made available through this library.

Re-exports

pub use external_deps::c_build;

pub use external_deps::c_cxx_compiler;

pub use external_deps::clang;

pub use external_deps::htmldocck;

pub use external_deps::llvm;

pub use external_deps::python;

pub use external_deps::rustc;

pub use external_deps::rustdoc;

pub use c_cxx_compiler::Cc;

pub use c_cxx_compiler::Gcc;

pub use c_cxx_compiler::cc;

pub use c_cxx_compiler::cxx;

pub use c_cxx_compiler::extra_c_flags;

pub use c_cxx_compiler::extra_cxx_flags;

pub use c_cxx_compiler::gcc;

pub use c_build::build_native_dynamic_lib;

pub use c_build::build_native_static_lib;

pub use c_build::build_native_static_lib_cxx;

pub use c_build::build_native_static_lib_optimized;

pub use cargo::cargo;

pub use clang::clang;

pub use clang::Clang;

pub use htmldocck::htmldocck;

pub use llvm::llvm_ar;

pub use llvm::llvm_bcanalyzer;

pub use llvm::llvm_dis;

pub use llvm::llvm_dwarfdump;

pub use llvm::llvm_filecheck;

pub use llvm::llvm_nm;

pub use llvm::llvm_objcopy;

pub use llvm::llvm_objdump;

pub use llvm::llvm_profdata;

pub use llvm::llvm_readobj;

pub use llvm::LlvmAr;

pub use llvm::LlvmBcanalyzer;

pub use llvm::LlvmDis;

pub use llvm::LlvmDwarfdump;

pub use llvm::LlvmFilecheck;

pub use llvm::LlvmNm;

pub use llvm::LlvmObjcopy;

pub use llvm::LlvmObjdump;

pub use llvm::LlvmProfdata;

pub use llvm::LlvmReadobj;

pub use python::python_command;

pub use rustc::aux_build;

pub use rustc::bare_rustc;

pub use rustc::rustc;

pub use rustc::rustc_path;

pub use rustc::Rustc;

pub use rustdoc::bare_rustdoc;

pub use rustdoc::rustdoc;

pub use rustdoc::Rustdoc;

pub use diff::diff;

pub use diff::Diff;

pub use env::env_var;

pub use env::env_var_os;

pub use env::set_current_dir;

pub use run::cmd;

pub use run::run;

pub use run::run_fail;

pub use run::run_with_args;

pub use targets::is_aix;

pub use targets::is_darwin;

pub use targets::is_msvc;

pub use targets::is_windows;

pub use targets::llvm_components_contain;

pub use targets::target;

pub use targets::uname;

pub use targets::apple_os;

pub use artifact_names::bin_name;

pub use artifact_names::dynamic_lib_extension;

pub use artifact_names::dynamic_lib_name;

pub use artifact_names::msvc_import_dynamic_lib_name;

pub use artifact_names::rust_lib_name;

pub use artifact_names::static_lib_name;

pub use path_helpers::cwd;

pub use path_helpers::filename_contains;

pub use path_helpers::filename_not_in_denylist;

pub use path_helpers::has_extension;

pub use path_helpers::has_prefix;

pub use path_helpers::has_suffix;

pub use path_helpers::not_contains;

pub use path_helpers::path;

pub use path_helpers::shallow_find_files;

pub use path_helpers::source_root;

pub use scoped_run::run_in_tmpdir;

pub use scoped_run::test_while_readonly;

pub use assertion_helpers::assert_contains;

pub use assertion_helpers::assert_contains_regex;

pub use assertion_helpers::assert_count_is;

pub use assertion_helpers::assert_dirs_are_equal;

pub use assertion_helpers::assert_equals;

pub use assertion_helpers::assert_not_contains;

pub use assertion_helpers::assert_not_contains_regex;

pub use string::count_regex_matches_in_files_with_extension;

pub use string::invalid_utf8_contains;

pub use string::invalid_utf8_not_contains;

pub use bstr;

pub use gimli;

pub use libc;

pub use object;

pub use regex;

pub use serde_json;

pub use similar;

pub use wasmparser;

Modules

artifact_names

A collection of helpers to construct artifact names, such as names of dynamic or static librarys which are target-dependent.

assertion_helpers

Collection of assertions and assertion-related helpers.

command 🔒

diff

env

external_deps

This module contains external tool dependencies that we assume are available in the environment, such as cc or python.

fs 🔒

macros 🔒

path_helpers

Collection of path-related helpers.

rfs

std::fs wrappers and assorted filesystem-related helpers. Public to tests as rfs to not be confused with std::fs.

run

scoped_run

Collection of helpers that try to maintain certain properties while running a test closure.

string

symbols

targets

util 🔒