run_make_support
Module artifact_names

A collection of helpers to construct artifact names, such as names of dynamic or static librarys which are target-dependent.

Functions

bin_name

Construct the binary (executable) name based on the target.

dynamic_lib_extension

Construct the dynamic library extension based on the target.

dynamic_lib_name

Construct the dynamic library name based on the target.

msvc_import_dynamic_lib_name

Construct the name of the import library for the dynamic library, exclusive to MSVC and accepted by link.exe.

rust_lib_name

Construct the name of a rust library (rlib).

static_lib_name

Construct the static library name based on the target.