run_make_support::external_deps
Module rustc

Structs

Rustc

A rustc invocation builder.

Functions

aux_build

Construct a new rustc aux-build invocation.

bare_rustc

Construct a plain rustc invocation with no flags set. Note that set_host_rpath still presets the environment variable HOST_RPATH_DIR by default.

rustc

Construct a new rustc invocation. This will automatically set the library search path as -L cwd(). Use bare_rustc to avoid this.

rustc_path

setup_common 🔒